fixed some bugs in report
