+++ transforms/code/higher_order_syntactic_profiler/README.md
@@ -1,6 +1,7 @@
-# NOOP Transform 
-The NOOP transforms serves as a simple exemplar to demonstrate the development
-of a simple 1:1 transform.  Per the set of 
+# SP Transform 
+The HOSP transform enables the profiling of a given code dataset based on higher order
+syntactic and semantic concepts as specified by the user. It also generates a profiling
+report in HTML based on the output table. Per the set of 
 [transform project conventions](../../README.md#transform-project-conventions)
 the following runtimes are available:
