+++ transforms/code/higher_order_syntactic_profiler/python/src/template.html
@@ -8,27 +8,54 @@
     <script src="https://cdn.plot.ly/plotly-latest.min.js"></script> <!-- Include Plotly JS -->
     <style>
         body {
-            font-size: 14px; /* Decrease the overall font size */
+            font-size: 14px;
         }
 
         h1.title {
-            font-size: 20px; /* Smaller title font */
+            font-size: 20px;
         }
 
         h2.title.is-4 {
-            font-size: 16px; /* Smaller metric title */
+            font-size: 16px;
         }
 
-        p, li {
-            font-size: 12px; /* Smaller text for details and lists */
+        p,
+        li {
+            font-size: 12px;
         }
 
         .box {
-            padding: 10px; /* Smaller padding for the metric box */
+            padding: 10px;
         }
 
         .metric {
-            margin-bottom: 10px; /* Reduce spacing between metrics */
+            margin-bottom: 10px;
+        }
+
+        .metrics-list {
+            border-right: 1px solid #ddd;
+            padding-right: 20px;
+        }
+
+        .metrics-container {
+            padding-left: 20px;
+        }
+
+        .columns {
+            display: flex;
+        }
+
+        .column {
+            overflow-y: auto;
+            max-height: 80vh;
+        }
+
+        .metrics-column {
+            flex: 1;
+        }
+
+        .plots-column {
+            flex: 3;
         }
     </style>
 </head>
@@ -40,23 +67,31 @@
             <p>{{ description }}</p>
             <hr />
 
-            <h3 class="subtitle">Available Metrics</h3>
-            <ul>
-                {% for metric in metrics %}
-                <li><a onclick="toggleVisibility('{{ metric.id }}')">{{ metric.name }}</a></li>
-                {% endfor %}
-            </ul>
-
-            <div id="metrics-container">
-                {% for metric in metrics %}
-                <div id="{{ metric.id }}" class="box metric" style="display: none;">
-                    <h2 class="title is-4">{{ metric.name }}</h2>
-                    <p>{{ metric.details }}</p>
-                    {% if metric.graph_html %}
-                    {{ metric.graph_html | safe }} <!-- Render the Plotly graph HTML -->
-                    {% endif %}
+            <div class="columns">
+                <!-- Metrics List on the Left -->
+                <div class="column metrics-column metrics-list">
+                    <h3 class="subtitle">Available Metrics</h3>
+                    <ul>
+                        {% for metric in metrics %}
+                        <li><a onclick="toggleVisibility('{{ metric.id }}')">{{ metric.name }}</a></li>
+                        {% endfor %}
+                    </ul>
                 </div>
-                {% endfor %}
+
+                <!-- Plot Frames on the Right -->
+                <div class="column plots-column metrics-container">
+                    <div id="metrics-container">
+                        {% for metric in metrics %}
+                        <div id="{{ metric.id }}" class="box metric" style="display: none;">
+                            <h2 class="title is-4">{{ metric.name }}</h2>
+                            <p>{{ metric.details }}</p>
+                            {% if metric.graph_html %}
+                            {{ metric.graph_html | safe }} <!-- Render the Plotly graph HTML -->
+                            {% endif %}
+                        </div>
+                        {% endfor %}
+                    </div>
+                </div>
             </div>
         </div>
     </section>
