+++ transforms/code/higher_order_syntactic_profiler/ray/README.md
@@ -1,4 +1,4 @@
-# NOOP Ray Transform 
+# SP Ray Transform 
 Please see the set of
 [transform project conventions](../../../README.md#transform-project-conventions)
 for details on general project conventions, transform configuration,
