--- transforms/code/higher_order_syntactic_profiler/ray/src/concept_map/updated_concept_list.csv
@@ -1,14 +1,0 @@
-Category
-Algorithms and Data Structures
-Database Management
-File Handling
-Networking and Messaging
-Graphical User Interface Design
-Security
-Scheduling and Concurrency
-Logging and Monitoring
-Web Development
-Mathematics and Numerics
-Code Analysis and Linting
-Testing
-Data Serialization
