--- transforms/code/higher_order_syntactic_profiler/ray/src/examples/examples-i.csv
@@ -1,27 +1,0 @@
-Library,Language
-algorithms,Python
-asyncio,Python
-arrow,Python
-authlib,Python
-webassets,Python
-scipy,Python
-pymysql,Python
-mimetypes,Python
-logging,Python
-flake8,Python
-mamba,Python
-marshmallow,Python
-tkinter,Python
-com.leansoft.bigqueue,Java
-com.cisco.commons.networking,Java
-net.time4j,Java
-org.apache.shiro,Java
-java.net.http,Java
-org.apache.commons.math4,Java
-ch.vorburger.mariaDB4j,Java
-com.google.jimfs,Java
-java.logging,Java
-org.sonar,Java
-org.junit,Java
-com.cedarsoftware:json-io,Java
-java.desktop,Java
