--- transforms/code/higher_order_syntactic_profiler/ray/src/ikb/null_libs.csv
@@ -1,14 +1,0 @@
-Library,Language
-dynamic_bitset,Cpp
-tries,Cpp
-algorithm,Cpp
-uni-algo,Cpp
-boost.asio,Cpp
-cpp-netlib,Cpp
-zmq,Cpp
-azmq,Cpp
-thread-pool,Cpp
-chrono,Cpp
-concurrencpp,Cpp
-time,Cpp
-libressl,Cpp
