+++ transforms/code/semantic_profiler/python/README.md
@@ -1,4 +1,4 @@
-# NOOP Transform 
+# SP Transform 
 Please see the set of
 [transform project conventions](../../../README.md#transform-project-conventions)
 for details on general project conventions, transform configuration,
@@ -14,14 +14,11 @@ of the transform is implemented.
 
 ## Configuration and command line Options
 
-The set of dictionary keys holding [NOOPTransform](src/noop_transform.py) 
+The set of dictionary keys holding [SPTransform](src/sp_transform.py) 
 configuration for values are as follows:
 
-* _noop_sleep_sec_ - specifies the number of seconds to sleep during the call the 
-the `transform()` method of `NOOPTransformation`.  This may be useful for
-simulating real transform timings and as a way to limit I/O on an S3 endpoint..
-* _noop_pwd_ - specifies a dummy password not included in metadata. Provided
-as an example of metadata that we want to not include in logging.
+* _sp_ikb_file_ - 
+* _sp_null_libs_file_ - 
 
 ## Running
 
@@ -30,17 +27,18 @@ The following command line arguments are available in addition to
 the options provided by 
 the [python launcher](../../../../data-processing-lib/doc/python-launcher-options.md).
 ```
-  --noop_sleep_sec NOOP_SLEEP_SEC
-                        Sleep actor for a number of seconds while processing the data frame, before writing the file to COS
-  --noop_pwd NOOP_PWD   A dummy password which should be filtered out of the metadata
+  --sp_ikb_file        SP_IKB_FILE
+                       Sleep actor for a number of seconds while processing the data frame, before writing the file to COS
+  --sp_null_libs_file  SP_NULL_LIBS_FILE   
+                       A dummy password which should be filtered out of the metadata
 ```
 These correspond to the configuration keys described above.
 
 ### Running the samples
 To run the samples, use the following `make` targets
 
-* `run-cli-sample` - runs src/noop_transform.py using command line args
-* `run-local-sample` - runs src/noop_local.py
+* `run-cli-sample` - runs src/sp_transform.py using command line args
+* `run-local-sample` - runs src/sp_local.py
 
 These targets will activate the virtual environment and set up any configuration needed.
 Use the `-n` option of `make` to see the detail of what is done to run the sample.
